decentralize default constants
