+++ .projenrc.ts
@@ -47,7 +47,7 @@ project.tryFindObjectFile('.github/workflows/build.yml')!.addOverride('jobs.buil
   CDK_DEFAULT_EMAIL: '${{ secrets.CDK_DEFAULT_EMAIL }}',
   TOWER_ACCESS_TOKEN: '${{ secrets.TOWER_ACCESS_TOKEN }}',
   TOWER_COMPUTE_ENV_ID: '${{ secrets.TOWER_COMPUTE_ENV_ID }}',
-  TOWER_OUTPUT_BUCKET: '${{ secrets.TOWER_OUTPUT_BUCKET }}',
+  CDK_DEFAULT_BUCKET: '${{ secrets.CDK_DEFAULT_BUCKET }}',
   TOWER_ORG: '${{ secrets.TOWER_ORG }}',
   TOWER_WORKSPACE: '${{ secrets.TOWER_WORKSPACE }}',
   TOWER_WORKSPACE_ID: '${{ secrets.TOWER_WORKSPACE_ID }}',
